Move setup and teardown into core
