+++ tests/test_replication_key.py
@@ -1,12 +1,8 @@
 """Tests standard tap features using the built-in SDK tests library."""
-import datetime
 import json
 
 import pendulum
-import sqlalchemy
-from faker import Faker
 from singer_sdk.testing.templates import TapTestTemplate
-from sqlalchemy import Column, DateTime, Integer, MetaData, String, Table
 
 from tap_postgres.tap import TapPostgres
 
@@ -17,45 +13,10 @@ SAMPLE_CONFIG = {
 }
 
 
-def setup_test_table(table_name, sqlalchemy_url):
-    """setup any state specific to the execution of the given module."""
-    engine = sqlalchemy.create_engine(sqlalchemy_url)
-    fake = Faker()
-
-    date1 = datetime.date(2022, 11, 1)
-    date2 = datetime.date(2022, 11, 30)
-    metadata_obj = MetaData()
-    test_replication_key_table = Table(
-        table_name,
-        metadata_obj,
-        Column("id", Integer, primary_key=True),
-        Column("updated_at", DateTime(), nullable=False),
-        Column("name", String()),
-    )
-    with engine.connect() as conn:
-        metadata_obj.create_all(conn)
-        conn.execute(f"TRUNCATE TABLE {table_name}")
-        for _ in range(1000):
-            insert = test_replication_key_table.insert().values(
-                updated_at=fake.date_between(date1, date2), name=fake.name()
-            )
-            conn.execute(insert)
-
-
-def teardown_test_table(table_name, sqlalchemy_url):
-    engine = sqlalchemy.create_engine(sqlalchemy_url)
-    with engine.connect() as conn:
-        conn.execute(f"DROP TABLE {table_name}")
-
-
 def replication_key_test(tap, table_name):
     """Originally built to address
     https://github.com/MeltanoLabs/tap-postgres/issues/9
     """
-    # Create sink for table
-    # tap = TapPostgres(
-    #     config=SAMPLE_CONFIG
-    # )  # If first time running run_discovery won't see the new table
     tap.run_discovery()
     # TODO Switch this to using Catalog from _singerlib as it makes iterating
     # over this stuff easier
